refactor: move API response contracts to root responses module
