+++ india_stocks_api/responses.py
