+++ tests/integration/test_data_methods_live.py
@@ -18,7 +18,7 @@ import pytest
 from india_stocks_api.brokers.angel import AngelOne
 from india_stocks_api.constants import CandleInterval
 from india_stocks_api.instruments import Equity
-from india_stocks_api.models import DepthResponse, FundsResponse, HistoryResponse, ProfileResponse, QuoteResponse
+from india_stocks_api.responses import DepthResponse, FundsResponse, HistoryResponse, ProfileResponse, QuoteResponse
 
 
 def _get_creds():
